Refactor: Rename Policy Store to Policy Model Store
Renames the SQLite-backed policy-model service from aiac-policy-store to
aiac-policy-model-store across code, tests, manifests, Docker image, and
demo/integration targets, freeing the old name/key/filename for ChromaDB
in Handoff 04.

- Python package aiac.policy.store -> aiac.policy.model_store (+ tests)
- Env key AIAC_POLICY_STORE_URL -> AIAC_POLICY_MODEL_STORE_URL
- k8s manifest policy-store-statefulset.yaml -> policy-model-store-statefulset.yaml
  (all identifiers; SERVICEPOLICY_DB_PATH and securityContext preserved)
- Image aiac-policy-store -> aiac-policy-model-store; Dockerfile moves with package
- Demo/integration svc target aiac-policy-model-store-service
- PRB import-isolation FORBIDDEN guard repointed to aiac.policy.model_store.library
